fix some dependencies
