+++ components/common/CharmEditor/components/emojiSuggest/emojiSuggest.constants.ts
@@ -1,0 +1,4 @@
+import { PluginKey } from '@bangle.dev/core';
+
+export const pluginKey = new PluginKey('emojiSuggestKey');
+export const markName = 'emojiSuggest';
