+++ components/common/CharmEditor/components/emojiSuggest/index.ts
@@ -1,0 +1,6 @@
+import EmojiSuggest from './EmojiSuggest.component';
+
+export * from './emojiSuggest.specs';
+export * from './emojiSuggest.plugins';
+
+export default EmojiSuggest;
