+++ components/common/CharmEditor/components/nestedPage/index.ts
@@ -1,0 +1,7 @@
+import NestedPage from './NestedPage.component';
+import NestedPagesList from './NestedPagesList';
+
+export * from './nestedPage';
+export { NestedPagesList };
+
+export default NestedPage;
